requirements.txt
google-generativeai>=0.3.0,<1
openai>=1.23.0
PyYAML>=6.0
python-dotenv>=1.0.0
loguru>=0.7.2
networkx>=3.0
fastapi>=0.109.0
uvicorn>=0.27.0